requirements.txt
Django==4.2.24
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.7
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.4
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.0
python-decouple==3.8
django-cors-headers==4.3.1
celery==5.3.1
redis==4.6.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
PyYAML==6.0.1
requests==2.31.0
psycopg2-binary==2.9.7
whitenoise==6.5.0